02_streamlit_pdf_chatbot/requirements.txt
streamlit
openai
python-dotenv
faiss-cpu
nltk
PyMuPDF
